Added Controllable Skateboard
The skateboard can not perform tricks through the Skateboard Controller Class, which responds to events triggered by by the Touch Controls Class.
The Touch Controls class stores the amount of time each touch/swipe took which is used to determine the multiplier added on the base jump force that is applied for each trick that it performed.