Added Obstacle ScriptableObjects
The Procedural Map Generator will spawn objects based on type, and then it will randomly choose a different object from the type chosen. For example, if bench is chosen (at random), it then will see how many benches can be spawned, and will chose one. It will also determine if another object should be spawner after, based on type as well. For example, it will not spawn an object after a curb, it will skip on sequence, and after that it will spawn again.