Game Speed checking on Obstacle Generation #30
The optimised version of obstacle generation will only spawn an obstacle if the current game speed matches its minimum acceptance speed. This is done by ordering the list of obstacles before creating the pool of obstacles, and find the last indexes of when the slow, medium and fast obstacles are in the list. The Pools GetRandomObject now takes a maxExclusive index to know when to stop searching for a random obstacle. This allows me to pass to that function the last slow index, the last medium index or the last fast index, essential ensuring that no obtacle is spawned when the right speed hasn't been reached.