Fixed Audio Lag Spike #42 #40
The sound playing on the Audio Manager cannot be optimised more. The lag spike occurred because the engine had to load multiple sounds at a time and then play them. Afterwards, no lag spike was caused because the sounds had already been loaded once. To fix this, we pre load the audio data of specific sounds that we are planning to play in conjuction to others, and so the engine doesn't have to load them first before playing them. Check Unity Docs here: https://docs.unity3d.com/ScriptReference/AudioClip-preloadAudioData.html